requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
apscheduler==3.10.4
jinja2==3.1.3
python-multipart==0.0.6
pyyaml==6.0.1
aiosqlite==0.19.0
